requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-binance>=1.0.19
ta>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
twelvedata>=1.2.0
